pyproject.toml
[project]
name = "mcp-server-redcap"
version = "0.2.0"
description = "MCP server for REDCap (Research Electronic Data Capture)"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Miguel-Angel Sicilia", email = "msicilia@gmail.com" },
]
keywords = ["mcp", "redcap", "research", "clinical-data", "ehr", "model-context-protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "PyCap>=2.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
]

[project.urls]
Homepage = "https://github.com/msicilia/mcp-server-redcap"
Repository = "https://github.com/msicilia/mcp-server-redcap"
Issues = "https://github.com/msicilia/mcp-server-redcap/issues"

[project.scripts]
mcp-server-redcap = "mcp_server_redcap.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_redcap"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
